The zip task will preserve the folder structure by default.
This is a change from previous behavior. The zip task used to flatten all
zips by default. And you had to set a bool property to preserve the folder
structure. Now, you have to call a void method, flatten_zip, to cause the
flattening behavior.

The whole behavior is rather odd, anyway, depending on if you're zipping
one or many directories. I hope this doesn't break too many builds. We'll
make a note in the final release notes.