Update vb.rb
This element is upper-cased, by default, in VB. Without a case-insensitive search, you'll generate an extra element and fail to compile.

This "fix" will result in a non-standard, lower-cased `assembly` element, however. Because the base engine takes care of creating the element. And I don't feel like tackling that right now.

Maybe in another refactor :)