Update aspnetcompiler.rb
Removed commented out properties and replaced bool properties with methods (breaking). I couldn't find `format_path`, at first, but it's in the `SupportsLinuxEnvironment` module in `supportslinux.rb`.

I cleaned up the failure message and I like this one more... I think we ought to default to this in the `fail_with_message` method. I'm supposed to provide the task name, like "AspNetCompiler", somewhere in the class structure, I think. We could reuse that.