Fixed CrossPlatformCmd.which for (absolute) paths with spaces.
Added quotation for input argument to which/where, in order to handle paths with spaces (e.g. "C:\Program Files (x86)\MSBuild\12.0\Bin\MSBuild.exe").

Differentiated input argument to which/where for absolute paths -- the "path:file" arg. format for where on Windows is not supported by which on Unix (neither on OS X, nor on Ubuntu Linux).